Handle finish events on transactions returned from cache
`Mojo::Transaction` has a [`finish` event](https://docs.mojolicious.org/Mojo/Transaction#finish) which can be subscribed to to trigger a callback when the transaction completes.  Because `Mojo::UserAgent::Cached` creates a new transaction containing the cached response, such a callback is only called when content is fetched, not when returned from cache.

This commit fixes it so that the `finish` callback is called even for transactions returning cached responses.

This seems to be sufficient for my use case (which is [Mojo::UserAgent::Role::Queued](https://metacpan.org/pod/Mojo::UserAgent::Role::Queued)).  More work would be required to support other events, such as the `finish` events on `req` and `res`.